Move connection handlers to cluster_link.c with vtable callbacks
Move clusterReadHandler, clusterAcceptHandler, clusterLinkConnectHandler,
and clusterConnAcceptHandler from cluster_legacy.c to cluster_link.c.
These are now protocol-agnostic, dispatching to vtable callbacks for
protocol-specific behavior.

Add vtable callbacks:
- validateMessageHeader: validate header and return total message length
- processMessage: process a complete message from link->rcvbuf
- postConnect: called after outbound link connection is established

Move RCVBUF_MIN_READ_LEN and RCVBUF_MAX_PREALLOC defines to
cluster_link.h.

Change the static_assert for RCVBUF_MIN_READ_LEN to use <= instead
of == for robustness.

Change clusterLegacyValidateMessageHeader to return totlen (0 if
invalid) instead of a boolean.
