+++ src/cluster_legacy.c
@@ -427,8 +427,6 @@ static_assert(offsetof(clusterMsgHeader, notused2) == offsetof(clusterMsg, count
 #define CLUSTER_BROADCAST_ALL 0            /* All known instances. */
 #define CLUSTER_BROADCAST_LOCAL_REPLICAS 1 /* All replicas in my primary-replicas ring. */
 
-void clusterAcceptHandler(aeEventLoop *el, int fd, void *privdata, int mask);
-void clusterReadHandler(connection *conn);
 void clusterSendPing(clusterLink *link, int type);
 void clusterSendFail(char *nodename);
 void clusterSendFailoverAuthIfNeeded(clusterNode *node, clusterMsg *request);
@@ -485,8 +483,8 @@ static inline int myselfIsBestRankedReplica(void) {
 #define SLOT_WORD_OFFSET(w) ((w) << 3)
 
 #define RCVBUF_MIN_READ_LEN 14
-static_assert(offsetof(clusterMsg, type) + sizeof(uint16_t) == RCVBUF_MIN_READ_LEN,
-              "Incorrect length to read to identify type");
+static_assert(offsetof(clusterMsg, type) + sizeof(uint16_t) <= RCVBUF_MIN_READ_LEN,
+              "RCVBUF_MIN_READ_LEN too small to read message type");
 
 #define RCVBUF_MAX_PREALLOC (1 << 20) /* 1MB */
 
@@ -953,75 +951,6 @@ clusterMsgSendBlock *createClusterMsgSendBlock(int type, uint32_t msglen) {
 }
 
 
-static void clusterConnAcceptHandler(connection *conn) {
-    clusterLink *link;
-
-    if (connGetState(conn) != CONN_STATE_CONNECTED) {
-        serverLog(LL_VERBOSE, "Error accepting cluster node connection: %s", connGetLastError(conn));
-        connClose(conn);
-        return;
-    }
-
-    /* Create a link object we use to handle the connection.
-     * It gets passed to the readable handler when data is available.
-     * Initially the link->node pointer is set to NULL as we don't know
-     * which node is, but the right node is references once we know the
-     * node identity. */
-    link = createClusterLink(NULL);
-    link->conn = conn;
-    connSetPrivateData(conn, link);
-
-    /* Register read handler */
-    connSetReadHandler(conn, clusterReadHandler);
-}
-
-void clusterAcceptHandler(aeEventLoop *el, int fd, void *privdata, int mask) {
-    int cport, cfd;
-    int max = server.tls_cluster ? server.max_new_tls_conns_per_cycle : server.max_new_conns_per_cycle;
-    char cip[NET_IP_STR_LEN];
-    int require_auth = TLS_CLIENT_AUTH_YES;
-    UNUSED(el);
-    UNUSED(mask);
-    UNUSED(privdata);
-
-    /* If the server is starting up, don't accept cluster connections:
-     * UPDATE messages may interact with the database content. */
-    if (server.primary_host == NULL && server.loading) return;
-
-    while (max--) {
-        cfd = anetTcpAccept(server.neterr, fd, cip, sizeof(cip), &cport);
-        if (cfd == ANET_ERR) {
-            if (anetRetryAcceptOnError(errno)) continue;
-            if (errno != EWOULDBLOCK) serverLog(LL_VERBOSE, "Error accepting cluster node: %s", server.neterr);
-            return;
-        }
-
-        connection *conn = connCreateAccepted(connTypeOfCluster(), cfd, &require_auth);
-
-        /* Make sure connection is not in an error state */
-        if (connGetState(conn) != CONN_STATE_ACCEPTING) {
-            serverLog(LL_VERBOSE, "Error creating an accepting connection for cluster node: %s",
-                      connGetLastError(conn));
-            connClose(conn);
-            return;
-        }
-
-        connKeepAlive(conn, server.cluster_node_timeout / 1000 * 2);
-
-        /* Use non-blocking I/O for cluster messages. */
-        serverLog(LL_VERBOSE, "Accepting cluster node connection from %s:%d", cip, cport);
-
-        /* Accept the connection now.  connAccept() may call our handler directly
-         * or schedule it for later depending on connection implementation.
-         */
-        if (connAccept(conn, clusterConnAcceptHandler) == C_ERR) {
-            if (connGetState(conn) == CONN_STATE_ERROR)
-                serverLog(LL_VERBOSE, "Error accepting cluster node connection: %s", connGetLastError(conn));
-            connClose(conn);
-            return;
-        }
-    }
-}
 
 /* Return the approximated number of sockets we are using in order to
  * take the cluster bus connections. */
@@ -3453,20 +3382,8 @@ int clusterProcessPacket(clusterLink *link) {
 /* A connect handler that gets called when a connection to another node
  * gets established.
  */
-void clusterLinkConnectHandler(connection *conn) {
-    clusterLink *link = connGetPrivateData(conn);
+static void clusterLegacyPostConnect(clusterLink *link) {
     clusterNode *node = link->node;
-
-    /* Check if connection succeeded */
-    if (connGetState(conn) != CONN_STATE_CONNECTED) {
-        serverLog(LL_VERBOSE, "Connection with Node %.40s at %s:%d failed: %s", node->name, node->ip, node->cport,
-                  connGetLastError(conn));
-        freeClusterLink(link);
-        return;
-    }
-
-    /* Register a read handler from now on */
-    connSetReadHandler(conn, clusterReadHandler);
 
     /* Queue a PING in the new connection ASAP: this is crucial
      * to avoid false positives in failure detection.
@@ -3493,106 +3410,17 @@ void clusterLinkConnectHandler(connection *conn) {
      * the MEET flag here leads to asymmetry in the cluster membership. So, we
      * clear the MEET flag in clusterProcessPacket.
      */
-
-    serverLog(LL_DEBUG, "Connecting with Node %.40s at %s:%d", node->name, node->ip, node->cport);
 }
 
-/* Performs sanity check on the message signature and length depending on the type. */
-static inline int isClusterMsgSignatureAndLengthValid(clusterMsgHeader *hdr) {
+/* Validate the message header and return the total message length, or 0 if invalid. */
+static uint32_t clusterLegacyValidateMessageHeader(char *header) {
+    clusterMsgHeader *hdr = (clusterMsgHeader *)header;
     if (memcmp(hdr->sig, "RCmb", 4) != 0) return 0;
     uint16_t type = ntohs(hdr->type);
     uint32_t totlen = ntohl(hdr->totlen);
     uint32_t minlen = IS_LIGHT_MESSAGE(type) ? CLUSTERMSG_LIGHT_MIN_LEN : CLUSTERMSG_MIN_LEN;
     if (totlen < minlen) return 0;
-    return 1;
-}
-
-/* Read data. Try to read the first field of the header first to check the
- * full length of the packet. When a whole packet is in memory this function
- * will call the function to process the packet. And so forth. */
-void clusterReadHandler(connection *conn) {
-    clusterMsg buf[1];
-    ssize_t nread;
-    clusterMsgHeader *hdr;
-    clusterLink *link = connGetPrivateData(conn);
-    unsigned int readlen, rcvbuflen;
-
-    while (1) { /* Read as long as there is data to read. */
-        rcvbuflen = link->rcvbuf_len;
-        if (rcvbuflen < RCVBUF_MIN_READ_LEN) {
-            /* First, obtain the first 16 bytes to get the full message
-             * length and type. */
-            readlen = RCVBUF_MIN_READ_LEN - rcvbuflen;
-        } else {
-            /* Finally read the full message. */
-            hdr = (clusterMsgHeader *)link->rcvbuf;
-            if (rcvbuflen == RCVBUF_MIN_READ_LEN) {
-                /* Perform some sanity check on the message signature
-                 * and length. */
-                if (!isClusterMsgSignatureAndLengthValid(hdr)) {
-                    char ip[NET_IP_STR_LEN];
-                    int port;
-                    if (connAddrPeerName(conn, ip, sizeof(ip), &port) == -1) {
-                        serverLog(LL_WARNING, "Bad message length or signature received "
-                                              "on the Cluster bus.");
-                    } else {
-                        serverLog(LL_WARNING,
-                                  "Bad message length or signature received "
-                                  "on the Cluster bus from %s:%d",
-                                  ip, port);
-                    }
-                    freeClusterLink(link);
-                    return;
-                }
-            }
-            readlen = ntohl(hdr->totlen) - rcvbuflen;
-            if (readlen > sizeof(buf)) readlen = sizeof(buf);
-        }
-
-        nread = connRead(conn, buf, readlen);
-        if (nread == -1 && (connGetState(conn) == CONN_STATE_CONNECTED)) return; /* No more data ready. */
-
-        if (nread <= 0) {
-            /* I/O error... */
-            serverLog(LL_DEBUG, "I/O error reading from node link (%.40s:%s) (%s): %s",
-                      clusterLinkGetNodeName(link),
-                      link->inbound ? "inbound" : "outbound",
-                      clusterLinkGetHumanNodeName(link),
-                      (nread == 0) ? "connection closed" : connGetLastError(conn));
-            freeClusterLink(link);
-            return;
-        } else {
-            /* Read data and recast the pointer to the new buffer. */
-            size_t unused = link->rcvbuf_alloc - link->rcvbuf_len;
-            if ((size_t)nread > unused) {
-                size_t required = link->rcvbuf_len + nread;
-                size_t prev_rcvbuf_alloc = link->rcvbuf_alloc;
-                /* If less than 1mb, grow to twice the needed size, if larger grow by 1mb. */
-                link->rcvbuf_alloc = required < RCVBUF_MAX_PREALLOC ? required * 2 : required + RCVBUF_MAX_PREALLOC;
-                link->rcvbuf = zrealloc(link->rcvbuf, link->rcvbuf_alloc);
-                server.stat_cluster_links_memory += link->rcvbuf_alloc - prev_rcvbuf_alloc;
-            }
-            memcpy(link->rcvbuf + link->rcvbuf_len, buf, nread);
-            link->rcvbuf_len += nread;
-            hdr = (clusterMsgHeader *)link->rcvbuf;
-            rcvbuflen += nread;
-        }
-
-        /* Total length obtained? Process this packet. */
-        if (rcvbuflen >= RCVBUF_MIN_READ_LEN && rcvbuflen == ntohl(hdr->totlen)) {
-            if (clusterProcessPacket(link)) {
-                if (link->rcvbuf_alloc > RCVBUF_INIT_LEN) {
-                    size_t prev_rcvbuf_alloc = link->rcvbuf_alloc;
-                    zfree(link->rcvbuf);
-                    link->rcvbuf = zmalloc(link->rcvbuf_alloc = RCVBUF_INIT_LEN);
-                    server.stat_cluster_links_memory += link->rcvbuf_alloc - prev_rcvbuf_alloc;
-                }
-                link->rcvbuf_len = 0;
-            } else {
-                return; /* Link no longer valid. */
-            }
-        }
-    }
+    return totlen;
 }
 
 /* Put the message block into the link's send queue.
@@ -5784,6 +5612,9 @@ clusterBusType clusterLegacyBus = {
     .cron = clusterLegacyCron,
     .beforeSleep = clusterLegacyBeforeSleep,
     .handleServerShutdown = clusterLegacyHandleServerShutdown,
+    .validateMessageHeader = clusterLegacyValidateMessageHeader,
+    .processMessage = clusterProcessPacket,
+    .postConnect = clusterLegacyPostConnect,
     .updateMyselfFlags = clusterLegacyUpdateMyselfFlags,
     .updateMyselfIp = clusterLegacyUpdateMyselfIp,
     .updateMyselfHostname = clusterLegacyUpdateMyselfHostname,
