+++ src/cluster_link.h
@@ -7,6 +7,8 @@
 typedef struct clusterNode clusterNode;
 
 #define RCVBUF_INIT_LEN 1024
+#define RCVBUF_MIN_READ_LEN 14
+#define RCVBUF_MAX_PREALLOC (1 << 20) /* 1MB */
 
 /* A refcounted block of bytes queued for sending on a cluster link.
  * The data member is uint64_t to ensure alignment for protocol
@@ -48,6 +50,10 @@ char *clusterLinkGetHumanNodeName(clusterLink *link);
 int nodeIp2String(char *buf, clusterLink *link, char *announced_ip);
 void clusterLinkSendBlock(clusterLink *link, clusterMsgSendBlock *msgblock);
 int freeClusterLinkOnBufferLimitReached(clusterLink *link);
+
+void clusterReadHandler(connection *conn);
+void clusterLinkConnectHandler(connection *conn);
+void clusterAcceptHandler(aeEventLoop *el, int fd, void *privdata, int mask);
 
 struct client;
 void clusterCommandLinks(struct client *c);
